Move tests to correct package (#92)
