+++ src/test/scala/za/co/absa/hyperdrive/trigger/models/filters/DateTimeRangeFilterTest.scala
@@ -13,12 +13,11 @@
  * limitations under the License.
  */
 
-package za.co.absa.hyperdrive.trigger.filter
+package za.co.absa.hyperdrive.trigger.models.filters
 
 import java.time.LocalDateTime
 
 import org.scalatest.{FlatSpec, _}
-import za.co.absa.hyperdrive.trigger.models.filters.{DateTimeRangeFilter, DateTimeRangeFilterAttributes}
 
 class DateTimeRangeFilterTest extends FlatSpec with Matchers with BeforeAndAfterAll with BeforeAndAfterEach with FilterTestBase {
   import h2Profile.api._
