+++ src/test/scala/za/co/absa/hyperdrive/trigger/models/filters/FilterHelperTest.scala
@@ -13,13 +13,11 @@
  * limitations under the License.
  */
 
-package za.co.absa.hyperdrive.trigger.filter
+package za.co.absa.hyperdrive.trigger.models.filters
 
 import java.time.LocalDateTime
-
 import org.scalatest.{FlatSpec, _}
-import za.co.absa.hyperdrive.trigger.filters.FilterHelper
-import za.co.absa.hyperdrive.trigger.models.filters.{ContainsFilter, ContainsFilterAttributes, DateTimeRangeFilterAttributes, FilterAttributes, FilterSearchRequest, IntRangeFilterAttributes, StringEqualsFilterAttributes}
+
 class FilterHelperTest extends FlatSpec with Matchers with BeforeAndAfterAll with BeforeAndAfterEach with FilterTestBase {
   import h2Profile.api._
